Warning fixes. (#356)
* Use cast to fix warnings.

* Remove all empty definitions of portCLEAN_UP_TCB( pxTCB ) and
  portALLOCATE_SECURE_CONTEXT( ulSecureStackSize ) from ports.
  When these are undefined, the default empty definition is defined
  in FreeRTOS.h.

+++ timers.c
@@ -462,7 +462,7 @@
             }
             else
             {
-                pxTimer->ucStatus &= ~tmrSTATUS_IS_AUTORELOAD;
+                pxTimer->ucStatus &= ( ( uint8_t ) ~tmrSTATUS_IS_AUTORELOAD );
             }
         }
         taskEXIT_CRITICAL();
@@ -550,7 +550,7 @@
         }
         else
         {
-            pxTimer->ucStatus &= ~tmrSTATUS_IS_ACTIVE;
+            pxTimer->ucStatus &= ( ( uint8_t ) ~tmrSTATUS_IS_ACTIVE );
         }
 
         /* Call the timer callback. */
@@ -829,7 +829,7 @@
                             }
                             else
                             {
-                                pxTimer->ucStatus &= ~tmrSTATUS_IS_ACTIVE;
+                                pxTimer->ucStatus &= ( ( uint8_t ) ~tmrSTATUS_IS_ACTIVE );
                             }
 
                             /* Call the timer callback. */
@@ -846,7 +846,7 @@
                     case tmrCOMMAND_STOP:
                     case tmrCOMMAND_STOP_FROM_ISR:
                         /* The timer has already been removed from the active list. */
-                        pxTimer->ucStatus &= ~tmrSTATUS_IS_ACTIVE;
+                        pxTimer->ucStatus &= ( ( uint8_t ) ~tmrSTATUS_IS_ACTIVE );
                         break;
 
                     case tmrCOMMAND_CHANGE_PERIOD:
@@ -876,7 +876,7 @@
                                 }
                                 else
                                 {
-                                    pxTimer->ucStatus &= ~tmrSTATUS_IS_ACTIVE;
+                                    pxTimer->ucStatus &= ( ( uint8_t ) ~tmrSTATUS_IS_ACTIVE );
                                 }
                             }
                         #else /* if ( configSUPPORT_DYNAMIC_ALLOCATION == 1 ) */
@@ -885,7 +885,7 @@
                                  * could not have been dynamically allocated. So there is
                                  * no need to free the memory - just mark the timer as
                                  * "not active". */
-                                pxTimer->ucStatus &= ~tmrSTATUS_IS_ACTIVE;
+                                pxTimer->ucStatus &= ( ( uint8_t ) ~tmrSTATUS_IS_ACTIVE );
                             }
                         #endif /* configSUPPORT_DYNAMIC_ALLOCATION */
                         break;
